Repository Structure
Created a structure for our API
