+++ SummaryAPI/services/abstract.py
