+++ requirements.txt
@@ -1,2 +1,3 @@
+streamlit
 python-dotenv
 azure-ai-textanalytics==5.3.0b1
